Refactor for internal and external data sources
